+++ VfsInformationCustomerService 2010/VfsInformationCustomerSerive/DAOvEntitiesFramwork_CusServices/Report.cs
@@ -19,7 +19,7 @@ namespace DAOvEntitiesFramwork_CusServices
         public string UploadDir { get; set; }
         public Nullable<System.DateTime> DateViewCustomer { get; set; }
         public Nullable<long> TotalDownload { get; set; }
-        public Nullable<long> FileSize { get; set; }
+        public Nullable<int> FileSize { get; set; }
         public Nullable<int> IdReportType { get; set; }
         public string Ticker { get; set; }
         public System.DateTime CreateDate { get; set; }
